Update 0.9
Update 0.9 has introduced a whole new value (experimental) called secondaryType. secondaryType will allow for a type to be applied on top of a result of the base equation.

This type will have strings as shortcuts for some such as "squareroot" for Math.Sqr()

Options.js has also been made easier to read.