Made the calculator Take multiple Values
This required a lot of main.js restructuring to happen and there are still a lot of known issues that neead to be solved.

For example the only form of calculation at the moment is multiple and the old calculator has been commented out. This also leaves the MathTypes file unused. A lot of main.js error checking no longers works. The length checker and such have been removed.
String to number has been recoded... again. 

I still need to do a lot of stuff such as:
Add order of operations
Enable the old calculator if you wish
Honestly a lot of stuff where the entire internal structure of the calculator would need to be changed. May be best to have a separate branch and push the old version to a separate legacy branch.